requirements.txt
fastapi
openai
pydantic
uvicorn
python-dotenv
firebase_admin
pydantic-settings
PyPDF2
python-multipart
bleach
pydantic[email]
python-dateutil
pdfkit
python-docx
htmldocx
apscheduler